Outdoor mapping working with few exceptions.
Routing between buildings now works almost perfectly. The only exceptions
being that routing from a building to itself causes an error and if too
many requests are made to google within 10s the app will slow down until
google starts replying again.

+++ app/src/main/java/com/solutions/law/universityrouteplanner/FileReader.java
@@ -71,7 +71,7 @@ public class FileReader {
             line=bufferedReader.readLine();
             while(line!=null) {
                 linePieces = Arrays.asList(line.split(",")).iterator();
-                links.add(new Link(linePieces.next(),linePieces.next(),Integer.parseInt(linePieces.next())));
+                links.add(new Link(linePieces.next(),linePieces.next(),Double.parseDouble(linePieces.next())));
                 line = bufferedReader.readLine();
             }
         }catch (FileNotFoundException e){
